Updating the model to be more in-line with the closed source version of the model
The older model would have ESEa as a priority, with UKIC and events as the secondary. This was due to ensure incoming international players would still have some ranking points, in error this was not applied. This logic system now exists in the model. Coefficcients for the subsectors will likely be later updated though and this relationship further adjusted